Perf improvement - remove lots of Directory.Exists calls on the UI thread
In PR !17101 the HasExtension method was modified to call Directory.Exists which causes significant hammering on the filesystem APIs all over, especially on the UI thread. This change reverts the prior change and instead requires callers to correctly validate that they are working with files, which is actually easy to check without any file system APIs in almost all cases since they are generally working already on ListedItems or other objects that were loaded with that info. I use the Call Hierarchy tool to walk through all call paths to ensure I didn't miss any.