Perf improvement - Refactor StartSTATask to instead create an "STA thread pool"
StartSTATask was always creating a new thread for each task but this refactoring now caches a pool of threads created using DispatcherQueueController, which supports creating dedicated background STA threads. There is some perf win to not needing to create hundreds of threads for icon extraction and so forth. The implementation can be improved further.